Returns to old home page for continuous integration
